[hometask-5] Tests for light commands
