+++ src/test/java/ru/sbt/mipt/oop/CommandTests.java
@@ -1,39 +1,43 @@
 package ru.sbt.mipt.oop;
 
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-import ru.sbt.mipt.oop.commands.legacy.CommandType;
-import ru.sbt.mipt.oop.commands.legacy.SensorCommand;
-import ru.sbt.mipt.oop.commands.legacy.handlers.LightOffCommandHandler;
-import ru.sbt.mipt.oop.commands.legacy.handlers.SensorCommandHandler;
+import ru.sbt.mipt.oop.commands.CommandBuilder;
+import ru.sbt.mipt.oop.commands.CommandBuilderChain;
 
 public class CommandTests {
-    SmartHome smartHome;
+    private org.springframework.context.ApplicationContext context;
+    private SmartHome smartHome;
+    CommandBuilder commandBuilder;
 
     @Before
-    public void setUp() {
-        org.springframework.context.ApplicationContext context = new AnnotationConfigApplicationContext(TestConfiguration.class);
+    public void setUp() throws Exception {
+        context = new AnnotationConfigApplicationContext(TestConfiguration.class);
         smartHome = context.getBean(SmartHome.class);
+        commandBuilder = context.getBean(CommandBuilderChain.class).commandBuilder();
     }
 
     @Test
-    public void lightOffWorks() {
-        smartHome.execute(new LightCheck("2", true));
-        SensorCommandHandler handler = new LightOffCommandHandler(smartHome);
-        smartHome.execute(handler.handleCommand(
-                new SensorCommand(CommandType.LIGHT_OFF, "2")));
-        smartHome.execute(new LightCheck("2", false));
+    public void allRoomsLightOff() {
+        commandBuilder.build("allRoomsLightOff").execute();
+        smartHome.execute(component -> {
+            if (!(component instanceof Light))
+                return;
+            Light light = (Light) component;
+            Assert.assertFalse(light.isOn());
+        });
     }
 
     @Test
-    public void lightOffIsTargeted() {
-        smartHome.execute(new LightCheck("2", true));
-        smartHome.execute(new LightCheck("3", true));
-        SensorCommandHandler handler = new LightOffCommandHandler(smartHome);
-        smartHome.execute(handler.handleCommand(
-                new SensorCommand(CommandType.LIGHT_OFF, "2")));
-        smartHome.execute(new LightCheck("2", false));
-        smartHome.execute(new LightCheck("3", true));
+    public void allRoomsLightOn() {
+        commandBuilder.build("allRoomsLightOn").execute();
+        smartHome.execute(component -> {
+            if (!(component instanceof Light))
+                return;
+            Light light = (Light) component;
+            Assert.assertTrue(light.isOn());
+        });
     }
 }
